code adjustment for AvnuLatestAnalysisProvider
