+++ packages/plugin-starknet/src/providers/AvnuLatestAnalysisProvider.ts
@@ -1,5 +1,31 @@
 import { IAgentRuntime, Memory, Provider, State } from "@elizaos/core";
-import { fetchAvnuLatestAnalysis } from "./utils";
+import { AssetAnalysis } from "./types";
+
+const apiKey = process.env.BACKEND_API_KEY;
+
+const AVNU_ASSETS = [
+    'BROTHER', 'STRK'
+    , 'LORDS', 'USDC', 'ETH', 'UNI'
+    //'WBTC', 'UNI', 'RETH', 'XSTRK', 'NSTR',
+    //'ZEND', 'SWAY', 'SSTR'
+];
+
+const fetchAvnuLatestAnalysis = async (assets: string[] = AVNU_ASSETS): Promise<AssetAnalysis[]> => {
+    const assetsParam = assets.join(',');
+    const response = await fetch(
+        `http://127.0.0.1:8080/analysis/latest?assets=${assetsParam}&platform=avnu`,
+        {
+            headers: {
+                'accept': '*/*',
+                'x-api-key': apiKey,
+            }
+        }
+    );
+    
+    if (!response.ok) throw new Error(`HTTP Error: ${response.status}`);
+    return response.json();
+};
+
 
 const AvnuLatestAnalysisProvider: Provider = {
     get: async (_runtime: IAgentRuntime, _message: Memory, _state?: State) => {
